+++ grbl/stepper.c
@@ -143,7 +143,7 @@ typedef struct {
     uint8_t prescaler;      // Without AMASS, a prescaler is required to adjust for slow timing.
   #endif
   #ifdef VARIABLE_SPINDLE
-    uint8_t spindle_pwm;
+    uint32_t spindle_pwm;
   #endif
 } segment_t;
 static segment_t segment_buffer[SEGMENT_BUFFER_SIZE];
@@ -233,7 +233,7 @@ typedef struct {
 
   #ifdef VARIABLE_SPINDLE
     float inv_rate;    // Used by PWM laser mode to speed up segment calculations.
-    uint8_t current_spindle_pwm;
+    uint32_t current_spindle_pwm;
   #endif
 } st_prep_t;
 static st_prep_t prep;
@@ -548,7 +548,7 @@ void Timer1Proc()
       #endif
 
       #ifdef VARIABLE_SPINDLE
-        // Set real-time spindle output as segment is loaded, just prior to the first step.
+        // Set real-time spindle output as segment is loaded, just prior to the first step. CMMT 
         spindle_set_speed(st.exec_segment->spindle_pwm);
       #endif
